GHOST-607: Update rights on local clone before cleanup

Approved-by: Jérôme Respaut <shr3ps@gmail.com>
Approved-by: Kiruban PREMKUMAR <kpremkumar@morea.fr>
Approved-by: Burak Karamahmut <burak.karamahmut@fr.clara.net>
Approved-by: Patrick Decat <patrick.decat@fr.clara.net>
Approved-by: Quentin Manfroi <quentin.manfroi@yahoo.fr>

+++ commands/deploy.py
@@ -305,7 +305,9 @@ class Deploy():
         if self._is_commit_hash(revision):
             # Create intermediate clone from the local git mirror, chdir into it and fetch all commits
             source_path = get_intermediate_clone_path_from_module(self._app, module)
-            gcall('rm -rf {p}'.format(p=source_path), 'Removing previous intermediate clone', self._log_file)
+            if os.path.exists(source_path):
+                gcall('chmod -R u+rwx {p}'.format(p=source_path), 'Update rights on previous intermediate clone', self._log_file)
+                gcall('rm -rf {p}'.format(p=source_path), 'Removing previous intermediate clone', self._log_file)
             os.makedirs(source_path)
             os.chdir(source_path)
             gcall('du -hs .', 'Display current build directory disk usage', self._log_file)
@@ -319,7 +321,9 @@ class Deploy():
             gcall('du -hs .', 'Display current build directory disk usage', self._log_file)
 
             # Create shallow clone from the intermediate clone, chdir into it and retrieve submodules
-            gcall('rm -rf {p}'.format(p=clone_path), 'Removing previous clone', self._log_file)
+            if os.path.exists(clone_path):
+                gcall('chmod -R u+rwx {p}'.format(p=clone_path), 'Update rights on previous clone', self._log_file)
+                gcall('rm -rf {p}'.format(p=clone_path), 'Removing previous clone', self._log_file)
             os.makedirs(clone_path)
             os.chdir(clone_path)
             gcall('du -hs .', 'Display current build directory disk usage', self._log_file)
@@ -329,10 +333,13 @@ class Deploy():
             gcall('du -hs .', 'Display current build directory disk usage', self._log_file)
 
             # Destroy intermediate clone
+            gcall('chmod -R u+rwx {p}'.format(p=source_path), 'Update rights on previous intermediate clone', self._log_file)
             gcall('rm -rf {p}'.format(p=source_path), 'Removing intermediate clone', self._log_file)
         else:
             # Create clone from the local git mirror, chdir into it, fetch requested revision and retrieve submodules
-            gcall('rm -rf {p}'.format(p=clone_path), 'Removing previous clone', self._log_file)
+            if os.path.exists(clone_path):
+                gcall('chmod -R u+rwx {p}'.format(p=clone_path), 'Update rights on previous clone', self._log_file)
+                gcall('rm -rf {p}'.format(p=clone_path), 'Removing previous clone', self._log_file)
             os.makedirs(clone_path)
             os.chdir(clone_path)
             gcall('du -hs .', 'Display current build directory disk usage', self._log_file)
